backend/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
awscli==1.38.37
boto3==1.37.13
botocore==1.37.37
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.19
filelock==3.17.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
identify==2.6.7
idna==3.10
jmespath==1.0.1
nodeenv==1.9.1
oauthlib==3.2.2
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
s3transfer==0.11.4
sendgrid==6.11.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.2
wheel==0.45.1
